Move the alternative route search into HHAlternativeRoutes
The plateau method arrived as ~430 lines inside HHRoutePlanner, which already carries the
hub search, the region selection and the geometry expansion. Reviewing the two together is
hard and they change for different reasons, so the alternatives move to their own class.
It is a pure move: the code is unchanged, it calls back into the planner for the three
things only the planner can do (expand a final point, retrieve geometry, prepare results),
and those three methods lose "private" for that. What is left in HHRoutePlanner against
master is small enough to read at once: the search horizon in runRoutingWithInitQueue, the
acceptCostIncrease flag in retrieveSegmentsGeometry, and the call site.

Alternatives were also stored twice - as HHNetworkRouteRes in altRoutes and as plain
segment lists in RouteCalcResult.alternatives - which is what made append() clear both.
altRoutes is now the only storage and getAlternatives() is the same list seen through the
generic result, so nothing can go out of sync. RouteCalcResult keeps the accessor for
callers that do not know about HH and returns an empty list by default.

uniquePoints goes with the exclusion-ball code it belonged to: the old selection used it
to drop near-duplicate alternatives, the plateau method compares real road geometry
instead, and nothing has read the set since.
